admin_panel/requirements.txt
django==3.2
flake8==4.0
python-dotenv==1.0.0
psycopg2==2.9.6
django-debug-toolbar==3.4
split_settings==1.0.0
django-cors-headers==4.2.0
uwsgi==2.0.21
django-environ==0.10.0